flask_api/requirements.txt
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
transformers==4.44.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
accelerate==0.33.0
pytz==2024.1
bitsandbytes==0.43.3
huggingface-hub==0.24.5
haystack-ai==2.4.0
quanto==0.2.0
pgvector-haystack==0.5.1
pypdf==5.0.0
Flask==3.0.3
gunicorn==23.0.0
python-dotenv==1.0.1
pipdeptree==2.23.4
psycopg-binary==3.2.1